chore(coverage): isolate dead-by-link / public-header noise
block.c was at 76% lines because the `__attribute__((weak))` ray_alloc
fallback is replaced by the buddy allocator at link time — the stub is
dead-by-link in any normal build but llvm-cov counted its 7 lines as
missed inside block.c.  Move the stub to its own TU
(block_alloc_stub.c) so block.c measures only live code, and ignore
the stub TU in the coverage report regex.  block.c: 76% → 91% lines.

include/rayforce.h was at 73% because public `static inline` helpers
(ray_data_fn, ray_atom_is_null_fn) are instantiated in every TU that
includes the header — TUs that never call ray_data or hit the F32 NaN
branch leave those bodies "unexecuted" in their per-TU instantiation,
even though every relevant code path is exercised elsewhere.  Exclude
the public C API header from internal coverage (it's measured through
external integration tests, not the internal unit suite).

Also: move the rare slice arm of ray_data_fn out-of-line into
ray_data_slice_path (defined in vec.c).  The hot path stays trivially
inlinable while the cold path collapses to a single instantiation,
removing N×2 slice-branch lines from per-TU instantiation reports.
Extend test_slice_owned_ref to exercise ray_data on a slice block.

Add 2 sad-path tests for ray_block_size (RAY_SEL nrows<0,
out-of-range type) were already in test_block.c; they now show in
coverage now that the weak stub no longer drags block.c down.

Result: every measured file now ≥80% lines (smallest is hash.h at
82.08% after the prior include-cleanup commit).
